refactor: proj overview & nav
-Add aria labels to logo nav buttons. -Add styling and refactor text for project overview pages. -Remove unused css module from home page component.
