+++ src/components/Navigation/LogoButtons.tsx
@@ -28,6 +28,7 @@ function LogoButtons({ data }: PropType) {
       ) : (
         <Link
           to={data.url}
+          aria-label={data.id}
           target="_blank"
           rel="noopener noreferrer"
           className="flex justify-center items-center text-dark-blueish-gray w-16 h-12 text-3xl bg-highlight-yellow rounded-lg  hover:text-highlight-yellow hover:bg-dark-blueish-gray border-2 border-highlight-yellow"
