--- src/components/Pages/styles/Home.module.css
@@ -1,3 +1,0 @@
-.profile-img {
-  box-shadow: 0 0 8px yellow;
-}
