+++ src/data/ProjData copy.ts
@@ -39,7 +39,7 @@ const ProjData = [
       "Handled user authentication using Session Storage, social media authentication using the OAuth2 Framework, and wrote unit and integration tests for all components using Vitest to ensure robust security and reliable functionality.",
     ],
     challenges:
-      ["At the outset of this project, I encountered a significant challenge: while I was well-versed in JavaScript, React, Express, and PostgreSQL, I had no experience with some of the project requirements or technologies such as Prisma ORM or Remix, a server-side rendering JavaScript framework built on React. During the interview for this position, I was frank about my limitations, but was given a chance to hit the ground running. This presented a unique chance for me to step out of my comfort zone and quickly familiarize myself with new technologies on the job, all while ensuring timely delivery of project deliverables. Embracing this challenge head on has resulted in personal and professional growth as I have greatly expanded my skills in full-stack web application development.", "While I cannot disclose specific details about the hurdles I faced, this project was extensive and multifaceted. It involved developing interfaces for Landlord, Tenant, and Realtor accounts with complex integrations to facilitate effective communication and information sharing. Some technical challenges included debugging errors produced by code on the client, server, and external API, as well as designing database relationships and schema for proper information storage and flow. One crucial aspect involved ensuring that access to secure pages was restricted to authorized individuals based on account type. Navigating these challenges allowed me to further hone my programming skills, particularly in full-stack application development, problem-solving, and debugging. I am eager to apply these lessons to future projects and continue pushing the boundaries of what I can achieve as a software developer."],
+      ["At the outset of this project, I encountered a significant challenge: while I was well-versed in JavaScript, React, Express, and PostgreSQL, I had no experience with some of the project requirements or technologies such as Prisma ORM or Remix, a server-side rendering JavaScript framework built on React. During the interview for this position, I was frank about my limitations, but was given a chance to hit the ground running. This presented a unique obstacle that required me to step out of my comfort zone and quickly familiarize myself with new technologies. Embracing this challenge head on proved to be an exciting opportunity for personal and professional growth as I have greatly expanded my skills in full-stack web application development.", "While I cannot disclose specific details about the hurdles I faced, this project was extensive and multifaceted. It involved developing interfaces for Landlord, Tenant, and Realtor accounts with complex integrations to facilitate effective communication and information sharing. Some technical challenges included debugging errors produced by code on the client, server, and external API, as well as designing database relationships and schema for proper information storage and flow. One crucial aspect involved ensuring that access to secure pages was restricted to authorized individuals. Navigating these challenges allowed me to further hone my programming skills, particularly in problem-solving and debugging. I am eager to apply these lessons to future projects and continue pushing the boundaries of what I can achieve as a software developer."],
   },
   {
     id: uuidv4(),
@@ -78,7 +78,7 @@ const ProjData = [
       "Utilized Adobe Firefly AI and Chat GPT to programmatically generate image and JSON text content for the website.",
     ],
     challenges:
-      [""],
+      ["Biggest challenge of the project was getting the typing interface to work. -Loading too many characters at a time. -Deciding between allowing letters to overflow or break at the end of the line. -Finding a solution to have text disappear when user types more than two lines, and re-appear when user deletes all of the text available on screen. -Allowing for a responsive design without breaking the logic for displaying max two lines of text."],
   },
   {
     id: uuidv4(),
@@ -108,7 +108,7 @@ const ProjData = [
       "Tested, built, and maintained a mobile-responsive website using React, TypeScript, Tailwind CSS, Jest, and React Testing Library, to effectively highlight my personal and professional accomplishments as a software developer.",
     ],
     challenges:
-      [""],
+      ["Overhauled initial design that was very basic. -Move from using css modules to Tailwind CSS. Adding more skills required creating a hide/show menu while keeping the skills list responsive. "],
   },
   {
     id: uuidv4(),
@@ -142,7 +142,7 @@ const ProjData = [
       "Engineered streamlined design solutions for intricate challenges, fostering engagement with partners, investors, and developers, which, in turn, contributed to the enhancement of my communication and teamwork proficiencies.",
     ],
     challenges:
-      [""],
+      ["Collaborating with client, designers, and back-end team to provide desired results within a tight schedule. Simplifying the presentation of all information client wanted to showcase on the main landing page (show timeline, etc.). Researching and creating mockups, videos, and presentations on how to develop client's mobile/web app."],
   },
   {
     id: uuidv4(),
@@ -172,7 +172,7 @@ const ProjData = [
       "Managed domain, hosting, and site maintenance, ensuring uninterrupted user access and UX for nearly three years.",
     ],
     challenges:
-      [""],
+      ["Work closely with client to design and develop website based on their feedback. Placing a video on the landing page meant spending a lot of time optimizing it to improve page load speeds. Also creating the header was trick because the video had to be masked to provided the slanted design, then I created a drop-down list that overlays the video and also cycles through the list of services the company offers. Additionally, When the list is hovered, it opens up a drop down menu while simultaneously moving the call to action button out of the way to provide a better user experience. If I would do it over again, I would build it using React since it would be a lot easier to implement and manage compared to pure javascript, but it was a great learning experience that I thoroughly enjoyed."],
   },
 
   {
@@ -202,7 +202,7 @@ const ProjData = [
       "Programmed an Arduino Arc Fault detection system and mobile app utilizing MySQL for real-time notifications.",
     ],
     challenges:
-      [""],
+      ["Team member was not pulling their weight. As team lead I took the time to sit down and find out why. In the end I ended up taking on the coding work, as well as, managing the project to lessen the burden for the student. In the end our project was successful and the project group remains good friends to this day."],
   },
 ];
